Added au-extensions-1.2 draft-rfc
Added the template for the draft rfc and its relating xml example files.
Added a shell script to replace the example-file place holders with the
correct examples and generate the right xmls.
Added the main shell script for the project to replace tags in the
templates and build the html and txt format of the RFCs from the xml
generated.